Provide a base class HasDisplayTitle for classes providing a TITLE
Picard frequently uses the pattern were base classes can be extended by
plugins and offer a TITLE attribute for display purposes.

To make this TITLE translatable HasDisplayTitle can be used as a mixing.
It adds a display_title() class method that localizes the TITLE attribute,
if possible, and uses fallbacks to ensure there is a display title.